% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Extractor.R
\name{ArticleSentencesExtractor}
\alias{ArticleSentencesExtractor}
\title{A full-text extractor which is tuned towards extracting sentences from news articles.}
\usage{
ArticleSentencesExtractor(content, ...)
}
\arguments{
\item{content}{Text content as character}

\item{...}{additional parameters}
}
\value{
extracted text as character
}
\description{
A full-text extractor which is tuned towards extracting sentences from news articles.
}
\examples{
data(content)
extract <- ArticleSentencesExtractor(content)
}
\author{
Mario Annau
}
\seealso{
\code{\link{Extractor}}
}

