% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerpipeR-package.R
\docType{package}
\name{boilerpipeR-package}
\alias{boilerpipeR-package}
\alias{boilerpipe}
\title{Extract the main content from HTML files}
\description{
\pkg{boilerpipeR} interfaces the boilerpipe Java library, created by Christian
Kohlschutter \url{https://github.com/kohlschutter/boilerpipe}. It implements robust heuristics
to extract the main content from HTML files, removing unessecary
elements like ads, banners and headers/footers.
}
\examples{
\dontrun{
data(content)
extract <- DefaultExtractor(content)
cat(extract)
}
}
\seealso{
\code{\link{Extractor}} \code{\link{DefaultExtractor}} \code{\link{ArticleExtractor}}
}
\author{
Mario Annau \email{mario.annau@gmail}
}
\keyword{package}
