% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor.R
\name{DefaultExtractor}
\alias{DefaultExtractor}
\title{A quite generic full-text extractor.}
\usage{
DefaultExtractor(content, ...)
}
\arguments{
\item{content}{Text content as character}

\item{...}{additional parameters}
}
\value{
extracted text as character
}
\description{
A quite generic full-text extractor.
}
\examples{
data(content)
extract <- DefaultExtractor(content)
}
\seealso{
\code{\link{Extractor}}
}
\author{
Mario Annau
}
