% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor.R
\name{LargestContentExtractor}
\alias{LargestContentExtractor}
\title{A full-text extractor which extracts the largest text component of a page.}
\usage{
LargestContentExtractor(content, ...)
}
\arguments{
\item{content}{Text content as character}

\item{...}{additional parameters}
}
\value{
extracted text as character
}
\description{
For news articles, it may perform better than the \code{\link{DefaultExtractor}},
but usually worse than \code{\link{ArticleExtractor}}.
}
\examples{
data(content)
extract <- LargestContentExtractor(content)
}
\seealso{
\code{\link{Extractor}}
}
\author{
Mario Annau
}
