% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor.R
\name{ArticleExtractor}
\alias{ArticleExtractor}
\title{A full-text extractor which is tuned towards news articles.}
\usage{
ArticleExtractor(content, ...)
}
\arguments{
\item{content}{Text content as character}

\item{...}{additional parameters}
}
\value{
extracted text as character
}
\description{
In this scenario it achieves higher accuracy than \code{\link{DefaultExtractor}}.
}
\examples{
data(content)
extract <- ArticleExtractor(content)
}
\seealso{
\code{\link{Extractor}}
}
\author{
Mario Annau
}
