\name{ArticleExtractor}
\alias{ArticleExtractor}
\title{A full-text extractor which is tuned towards news articles.}
\usage{ArticleExtractor(content, ...)
}
\description{A full-text extractor which is tuned towards news articles. In this scenario
it achieves higher accuracy than \code{\link{DefaultExtractor}}.}
\arguments{\item{content}{Text content as character}
\item{...}{additional parameters}
}

