\name{ArticleExtractor}
\alias{ArticleExtractor}
\title{A full-text extractor which is tuned towards news articles. In this scenario
it achieves higher accuracy than \code{\link{DefaultExtractor}}.}
\usage{
  ArticleExtractor(content, ...)
}
\arguments{
  \item{content}{Text content as character}

  \item{...}{additional parameters}
}
\description{
  A full-text extractor which is tuned towards news
  articles. In this scenario it achieves higher accuracy
  than \code{\link{DefaultExtractor}}.
}

