% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor.R
\name{CanolaExtractor}
\alias{CanolaExtractor}
\title{A full-text extractor trained on a 'krdwrd' Canola (see \code{https://krdwrd.org/trac/attachment/wiki/Corpora/Canola/CANOLA.pdf}.}
\usage{
CanolaExtractor(content, ...)
}
\arguments{
\item{content}{Text content as character}

\item{...}{additional parameters}
}
\value{
extracted text as character
}
\description{
A full-text extractor trained on a 'krdwrd' Canola (see \code{https://krdwrd.org/trac/attachment/wiki/Corpora/Canola/CANOLA.pdf}.
}
\examples{
data(content)
extract <- CanolaExtractor(content)
}
\seealso{
\code{\link{Extractor}}
}
\author{
Mario Annau
}
