\name{CanolaExtractor}
\alias{CanolaExtractor}
\title{A full-text extractor trained on a...}
\usage{CanolaExtractor(content, ...)
}
\description{A full-text extractor trained on a
\href{http://krdwrd.org/}{krdwrd} \href{https://krdwrd.org/trac/attachment/wiki/Corpora/Canola/CANOLA.pdf}{Canola}.}
\arguments{\item{content}{Text content as character}
\item{...}{additional parameters}
}

