\docType{data}
\name{content}
\alias{content}
\title{This is data to be included in my package}
\description{
  This is data to be included in my package
}
\examples{
#Data set has been generated as follows:
\dontrun{
library(RCurl)
url <- "http://quantivity.wordpress.com/2012/11/09/multi-asset-market-regimes/"
content <- getURL(url)
}
}
\author{
  Mario Annau
}
\references{
  \url{http://quantivity.wordpress.com}
}
\keyword{data}
\keyword{datasets}

