\name{LargestContentExtractor}
\alias{LargestContentExtractor}
\title{A full-text extractor which extracts the largest text component of a page.}
\usage{LargestContentExtractor(content, ...)
}
\description{A full-text extractor which extracts the largest text component of a page.
For news articles, it may perform better than the \code{\link{DefaultExtractor}},
but usually worse than \code{\link{ArticleExtractor}}.}
\arguments{\item{content}{Text content as character}
\item{...}{additional parameters}
}

