\name{NumWordsRulesExtractor}
\alias{NumWordsRulesExtractor}
\title{A quite generic full-text extractor solely based upon the number of words per...}
\usage{NumWordsRulesExtractor(content, ...)
}
\description{A quite generic full-text extractor solely based upon the number of words per
block (the current, the previous and the next block).}
\arguments{\item{content}{Text content as character}
\item{...}{additional parameters}
}

